pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "staged-scraper"
version = "0.1.0"
description = "HTTP-first staged website-to-Markdown scraper with auditable escalation and deterministic Markdown output."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "OpenAI"}
]
dependencies = [
  "beautifulsoup4>=4.12,<5",
  "httpx>=0.28,<0.29",
  "lxml>=5,<7",
  "pydantic>=2.11,<3",
  "PyYAML>=6,<7",
  "rich>=14,<15",
  "typer>=0.21,<0.25",
]
keywords = ["scraping", "markdown", "crawler", "llms.txt", "playwright"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Text Processing :: Markup",
]

[project.optional-dependencies]
browser = [
  "playwright>=1.58,<2",
]
dev = [
  "pytest>=8,<10",
]

[project.scripts]
staged-scraper = "staged_scraper.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

testpaths = ["tests"]

[tool.ruff]
line-length = 180

target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]
